refactor(voice): centralize user_name encoding and chat message caps
- Move the pipe-delimited `user_name` layout into one client-safe module
  (src/lib/voice/user-name.ts) holding both buildUserName and a new
  parseUserName; mapParticipant and use-chat now decode via the shared
  parser instead of re-splitting on `|`. daily.ts re-exports both so the
  token routes import unchanged.
- Export a single MAX_MESSAGE_LENGTH from use-chat; ChatPanel imports it
  for the input maxLength so the cap and the receive-side trim can't drift.
- Drop unused sessionId/timestamp fields from ChatMessage.
